third_party/blink/web_tests/external/wpt/fetch/metadata/generated/fetch-via-serviceworker.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] sec-fetch-storage-access - Cross-site, init: mode=no-cors, credentials=include - respondWith
  assert_array_equals: expected property 0 to be "none" but got "active" (expected array ["none"] got ["active"])
[FAIL] sec-fetch-storage-access - Cross-site, init: mode=no-cors, credentials=include - fallback
  assert_array_equals: expected property 0 to be "none" but got "active" (expected array ["none"] got ["active"])
Harness: the test ran to completion.
